refactor: abstracted custom hooks for fetching user data, user listings (demo data)
